feat(account): 个人中心重构为统一账户 Hub，GitHub 档案降为子页
痛点：0.22.0 账户页把配额与 GitHub 个人主页杂糅在一条滚动流里（混乱），
纯邮箱/匿名用户界面太空。

重构为业界「我」Tab 范式的统一账户 Hub：
- 去掉 Home 顶栏齿轮，账户/头像入口铺到所有 tab 顶栏，未登录也可进；
- Hub 结构：身份区（含未登录引导）→ 套餐&用量卡（档位+余额进度+重置倒计时
  +费率+升级 CTA）→ GitHub 主页入口卡（仅 GitHub 用户）→ 设置分组（原设置页
  整体并入）→ 退出登录；
- GitHub 开发者档案（贡献图+动态流+下钻）拆为独立子页 GithubProfileScreen，
  与 Hub 共享同一 ProfileViewModel；
- 升级 CTA：Free→GitHub Sponsors / 匿名→登录引导 / Pro→致谢；后端零改动
  （复用 /api/me + /api/quota）；
- iOS（NoopAuthManager）自动退化为纯设置页。

实现：
- 新增 AccountScreen（Hub+并入设置）、GithubProfileScreen（GitHub 子页）；
  删除 ProfileScreen、SettingsScreen（内容拆分并入）。
- ProfileViewModel 支持匿名态：新增 loggedIn 字段；token 空不再判 error 而拉
  匿名档额度；loadQuota 仅在登录态 token 瞬态 null 时跳过；监听 authState 转变
  在登录/登出间实时切换。
- App.kt：Profile→AccountScreen，新增 GithubProfile 路由，移除 Settings 路由。
- HomeScreen：抽出 AccountAction 铺到三处 topbar，点头像直接进 Hub。

验证：Android+iOS 编译通过；shared 单测全绿（含 VM 匿名态回归）；Pixel_9_2
冒烟——匿名 Hub 渲染（5/5 匿名额度+设置分组全迁移）、登录选择器、跨 tab 入口
均正常。登录态 Hub 为旧 ProfileScreen 代码忠实抽离，待真机 OAuth 复验。
