+++ shared/src/commonMain/composeResources/values-zh/strings.xml
@@ -186,6 +186,19 @@
     <string name="profile_quota_reset_soon">1 小时内重置</string>
     <string name="profile_quota_rates">对话 1 · 搜索 3 · 研究 10</string>
     <string name="profile_quota_error">额度信息暂时无法获取</string>
+    <!-- 账户中心 -->
+    <string name="account_title">账户</string>
+    <string name="account_signed_out_title">未登录</string>
+    <string name="account_signed_out_prompt">登录以同步你的数据，并解锁更多每日额度</string>
+    <string name="account_plan_title">套餐与用量</string>
+    <string name="account_tier_free">免费版</string>
+    <string name="account_tier_pro">Pro</string>
+    <string name="account_upgrade_cta">升级 Pro</string>
+    <string name="account_upgrade_hint">Pro 权益绑定你赞助所用的 GitHub 账户。</string>
+    <string name="account_signin_for_more">登录即可获得每日 10 次额度</string>
+    <string name="account_pro_active">你已是 Pro 会员，感谢支持！</string>
+    <string name="account_github_entry">GitHub 主页</string>
+    <string name="account_github_entry_desc">贡献与关注动态</string>
     <string name="contrib_total">最近一年 %1$s 次贡献</string>
     <string name="contrib_less">少</string>
     <string name="contrib_more">多</string>
