+++ shared/src/commonTest/kotlin/whl/trending/ai/core/BackStackPopTest.kt
@@ -7,7 +7,7 @@ class BackStackPopTest {
 
     @Test
     fun pop_removes_top_entry() {
-        val backStack = mutableListOf<Any>(Home, Settings)
+        val backStack = mutableListOf<Any>(Home, Profile)
         backStack.safePop()
         assertEquals(listOf<Any>(Home), backStack)
     }
@@ -15,7 +15,7 @@ class BackStackPopTest {
     @Test
     fun double_pop_keeps_root() {
         // 模拟转场动画期间二次点击返回：第二次出栈必须被吞掉，栈底 Home 不可弹出
-        val backStack = mutableListOf<Any>(Home, Settings)
+        val backStack = mutableListOf<Any>(Home, Profile)
         backStack.safePop()
         backStack.safePop()
         assertEquals(listOf<Any>(Home), backStack)
@@ -30,8 +30,8 @@ class BackStackPopTest {
 
     @Test
     fun pop_deep_stack_removes_only_top() {
-        val backStack = mutableListOf<Any>(Home, Settings, Favorites, RepoDetail("a", "b"))
+        val backStack = mutableListOf<Any>(Home, Profile, Favorites, RepoDetail("a", "b"))
         backStack.safePop()
-        assertEquals(listOf<Any>(Home, Settings, Favorites), backStack)
+        assertEquals(listOf<Any>(Home, Profile, Favorites), backStack)
     }
 }
